requirements.txt
gradio>=4.44.0
openai>=1.40.0
python-dotenv>=1.0.1
requests>=2.32.0
huggingface_hub>=0.24.0
